refactor: 헤더 스크롤 관리하는 로직 훅으로 분리
